Update backend to work with json payloads.
According to the GCM documentation (here:
https://developers.google.com/cloud-messaging/http-server-ref#downstream-http-messages-plain-text),
plain text payloads have to be formatted as "data.<key>", but the
original simple-push-demo-server encoded this data by using

form_fields = {
    "registration_id": subscriptionId,
    "data": payload
}

urllib.urlencode(form_fields) did not generate the "data.<key>" keys for
GCM to read, so the expected intent extras are not passed to GCM on
Android.

Changing the server to send a JSON request (Content-Type:
application/json) and accepting JSON payloads will suit our needs in the
Advanced Android App Development quiz for the GCM fake server.